# Optimisation de l'indexing des documents Tentatives de batching et meilleur heuristique pour split by token len

+++ src/document_indexer.py
@@ -1,27 +1,29 @@
-# TODO : [ ] Upsert by batchs
 from langchain_mistralai import MistralAIEmbeddings
+import uuid
+
+from tqdm import tqdm
+from rich import print
 
 from src import config
-from qdrant_client.http.models import Distance, VectorParams, SparseVectorParams, Modifier
+from qdrant_client.http.models import Distance, VectorParams, SparseVectorParams, Modifier, PointStruct
 from langchain_text_splitters import RecursiveCharacterTextSplitter
 from langchain_qdrant import QdrantVectorStore, RetrievalMode, FastEmbedSparse
 from qdrant_client import QdrantClient
 
 from src.base_documentation_loader import BaseDocumentationLoader
 
-
 class DocumentIndexer:
-    def __init__(self, reset_collection = True):
+    def __init__(self, batch_size : int = 64, reset_collection = True):
         self.embedding = MistralAIEmbeddings(model=config.EMBEDDING_MODEL, api_key=config.MISTRAL_KEY)
         self.sparse_embedding = FastEmbedSparse(model_name=config.SPARSE_EMBEDDING_MODEL)
         self.client = QdrantClient(url=config.QDRANT_URL, api_key=config.QDRANT_KEY, port=None)
-
+        self.batch_size = batch_size
 
         self._setup_collection(reset_collection)
         self.vector_store = self._setup_vector_store()
 
     def _setup_collection(self, reset_collection):
-        if reset_collection :
+        if reset_collection and self.client.collection_exists(config.QDRANT_COLLECTION):
             self.client.delete_collection(config.QDRANT_COLLECTION)
 
         if self.client.collection_exists(config.QDRANT_COLLECTION):
@@ -46,6 +48,10 @@ class DocumentIndexer:
             sparse_vector_name=config.SPARSE_EMBEDDING_MODEL
         )
 
+    @staticmethod
+    def _token_len(text: str) -> int:
+        return int(len(text.split()) / 0.75)
+
     def index_documents(self, loader : BaseDocumentationLoader):
         print("Chargements des documents ...")
         docs = loader.load()
@@ -55,10 +61,18 @@ class DocumentIndexer:
         splitter = RecursiveCharacterTextSplitter(
             chunk_size=config.CHUNK_SIZE,
             chunk_overlap=config.CHUNK_OVERLAP,
-            length_function=len,
-            is_separator_regex=False,
+            length_function=self._token_len,
+            separators=["\n\n", "\n", " "],
         )
 
         chunks = splitter.split_documents(docs)
-        self.vector_store.add_documents(chunks)
-        print("Documents embedded")
+        print("Nombres de chunks créer :", len(chunks))
+        print(chunks[0])
+
+        # === Batching des chunks ===
+        batch_size = self.batch_size
+        for i in tqdm(range(0, len(chunks), batch_size), desc="Embedding batches"):
+            batch_chunks = chunks[i:(i + batch_size)]
+            self.vector_store.add_documents(batch_chunks)
+
+        print("\nDocuments embedded")
